refactor(ui): unify Health + Ops + Executions into single Operations nav area (#1109)
Collapse the three separate operational top-nav entries — Health
(/monitoring), Ops (/operating-room), and Executions (/executions) — into
a single Operations area at /operations: a ?tab=-driven tabbed view of
Needs Response · Notifications · Health · Executions · Resolved.

- Rename views/OperatingRoom.vue → Operations.vue (route /operations)
- Extract Health/Executions content into tab-embeddable
  components/MonitoringPanel.vue + ExecutionsPanel.vue; delete the
  standalone views/Monitoring.vue and views/Executions.vue
- Health tab admin-gated at the tab level (authStore.role); non-admin
  ?tab=health deep links coerced to the default tab
- Tabs toggle by v-if so each panel's store-owned polling tears down on
  leave; operator-queue polling stays container-level
- NavBar: single Operations link + one unified badge (operator-queue +
  notifications, critical-pulse); drop the separate Executions
  running-count badge (running count lives in the Executions tab strip)
- Function-form, query-preserving redirects: /monitoring→?tab=health,
  /executions→?tab=executions, /operating-room (preserve ?tab=),
  /events→?tab=notifications
- Per-execution detail route (ExecutionDetail.vue) unchanged
- Pure frontend IA change; no backend endpoints change
- Docs: architecture.md + operating-room / executions-dashboard /
  agent-monitoring / events-page feature-flows synced

Fixes #1109
