+++ docs/memory/architecture.md
@@ -285,6 +285,8 @@ Each agent runs as an isolated Docker container with standardized interfaces for
 - `stores/loops.js` - Sequential agent loops UI state, agent-scoped, WebSocket-driven live progress (#1106)
 - `stores/executions.js` - Fleet execution list/stats **+ agent Overview analytics** (`fetchAgentAnalytics`, cached per `${name}:${window}`, never polled) (#1107)
 
+**Top-nav IA — Operations (#1109):** the former separate **Health** (`/monitoring`), **Ops** (`/operating-room`), and **Executions** (`/executions`) top-nav entries are consolidated into a single **Operations** entry (`views/Operations.vue`, route `/operations`) — a `?tab=`-driven tabbed view: **Needs Response · Notifications · Health · Executions · Resolved**. Health/Executions content lives in tab-embeddable `components/MonitoringPanel.vue` / `ExecutionsPanel.vue` (extracted from the deleted standalone views, mirroring the #1107 `*Panel.vue` pattern); each tab is toggled by `v-if` so its store-owned polling tears down on tab-leave. The **Health** tab is admin-gated at the tab level (`authStore.role === 'admin'`) and non-admin `?tab=health` deep links are coerced to the default tab. The NavBar carries **one** unified badge (pending operator-queue + notifications, with critical-pulse) — the separate executions running-count badge is dropped (running count lives inside the Executions tab). Legacy `/monitoring`, `/executions`, `/operating-room`, `/events` routes redirect (function-form, query-preserving) to the matching Operations tab. Pure frontend IA change; no backend endpoints change. The per-execution detail route (`/agents/:name/executions/:executionId`, `ExecutionDetail.vue`) is unchanged.
+
 **Real-time:**
 - WebSocket client at `utils/websocket.js`
 - Auto-reconnect on disconnect
@@ -738,7 +740,7 @@ agent — detects the §P5 silent-clobber setup at fleet level.
 - `triggered_by` is grouped in Python via `_TRIGGER_BUCKETS` (`Chat/Tasks`, `MCP`, `Channels`, `Public`, `Scheduled`, `Agent-to-agent`, `Voice`) with an explicit **`Other`** catch-all so a new trigger type never silently vanishes. `manual` (the dominant interactive value) maps to `Chat/Tasks`.
 - `success_rate` is terminal-based — success / (success + failed [incl. `error`]); days with zero terminal rows report `success_rate=null` so the chart renders a gap, not a false 0%. `context_avg` uses NULL-skipping `AVG` so unmeasured rows don't read as 0.
 
-**Web UI:** a new **Overview** tab (`components/OverviewPanel.vue`) is the **default landing tab** on Agent Detail (first in `visibleTabs`; `activeTab` defaults to `'overview'`). It owns "trend over the last few days" while the persistent `AgentHeader` owns "now + cost" — Overview deliberately does **not** re-render the header's live CPU/MEM gauges, cost cards, git controls, autonomy/read-only/auth chips, or circuit badge (it links up to them). Sections: About lead (+ deep-link to Tasks/Info), a compact "needs attention" **count + link** to the Operating Room (hidden at zero), the trend charts, a health panel (current status + heartbeat + restart/OOM, with uptime/latency trend lines clamped to ≤7d by `agent_health_checks` retention and labeled "last 7 days"), a recent-activity drill-in, and footprint chips. Charts: `StackedBarChart.vue` (CSS/flexbox stacked bars — deliberately not uPlot bars, for correct-by-construction per-segment tooltips + theme colors) for executions-by-type, and `TrendLineChart.vue` (uPlot with axes/cursor, dark-mode-aware) for success-rate / duration / context / uptime / latency. **Info tab redesign:** `InfoPanel.vue` leads with the About narrative + "What You Can Ask" and tucks the exhaustive `template.yaml` metadata behind a collapsible "Technical details" `<details>` disclosure. Frontend Invariants #6/#7. (#1107)
+**Web UI:** a new **Overview** tab (`components/OverviewPanel.vue`) is the **default landing tab** on Agent Detail (first in `visibleTabs`; `activeTab` defaults to `'overview'`). It owns "trend over the last few days" while the persistent `AgentHeader` owns "now + cost" — Overview deliberately does **not** re-render the header's live CPU/MEM gauges, cost cards, git controls, autonomy/read-only/auth chips, or circuit badge (it links up to them). Sections: About lead (+ deep-link to Tasks/Info), a compact "needs attention" **count + link** to Operations (hidden at zero), the trend charts, a health panel (current status + heartbeat + restart/OOM, with uptime/latency trend lines clamped to ≤7d by `agent_health_checks` retention and labeled "last 7 days"), a recent-activity drill-in, and footprint chips. Charts: `StackedBarChart.vue` (CSS/flexbox stacked bars — deliberately not uPlot bars, for correct-by-construction per-segment tooltips + theme colors) for executions-by-type, and `TrendLineChart.vue` (uPlot with axes/cursor, dark-mode-aware) for success-rate / duration / context / uptime / latency. **Info tab redesign:** `InfoPanel.vue` leads with the About narrative + "What You Can Ask" and tucks the exhaustive `template.yaml` metadata behind a collapsible "Technical details" `<details>` disclosure. Frontend Invariants #6/#7. (#1107)
 
 ### Operator Queue (OPS-001)
