+++ docs/memory/feature-flows.md
@@ -11,6 +11,7 @@
 
 | Date | ID | Feature | Flow |
 |------|-----|---------|------|
+| 2026-06-09 | #1109 | refactor(ui): unify Health + Ops + Executions into a single **Operations** nav area — the three separate top-nav entries (Health `/monitoring`, Ops `/operating-room`, Executions `/executions`) collapse into one **Operations** link (`views/Operations.vue`, route `/operations`), a `?tab=`-driven 5-tab view: **Needs Response · Notifications · Health · Executions · Resolved**. `views/OperatingRoom.vue`→`Operations.vue`; Health/Executions content extracted into tab-embeddable `components/MonitoringPanel.vue`/`ExecutionsPanel.vue` (mirroring the #1107 `*Panel.vue` pattern), and the standalone `views/Monitoring.vue`/`views/Executions.vue` deleted. Tabs toggle by `v-if` so each panel's store-owned polling tears down on leave; operator-queue polling stays container-level. **Health** tab is admin-gated at the tab level (`authStore.role === 'admin'`) with non-admin `?tab=health` deep links coerced to the default tab. NavBar carries **one** unified badge (operator-queue + notifications, critical-pulse); the separate Executions running-count badge is dropped (running count lives in the Executions tab strip). Legacy `/monitoring`→`?tab=health`, `/executions`→`?tab=executions`, `/operating-room` (function-form, query-preserving), and `/events`→`?tab=notifications` all redirect. Per-execution detail route (`ExecutionDetail.vue`) unchanged. Pure frontend IA change; no backend endpoints change. | [operating-room.md](feature-flows/operating-room.md), [executions-dashboard.md](feature-flows/executions-dashboard.md), [agent-monitoring.md](feature-flows/agent-monitoring.md) |
 | 2026-06-09 | #1107 | feat(ui): Agent Detail **Overview** dashboard as the default landing tab + Info-tab redesign. New `OverviewPanel.vue` leads with About, a compact "needs attention" count+link, and **multi-day trend charts** (executions stacked by user-facing type bucket, terminal success-rate %, duration avg+p95, context) over a 7/14/30-day selector — plus a health panel (current status + heartbeat + restart/OOM; uptime/latency trend lines clamped ≤7d by `agent_health_checks` retention), a recent-activity drill-in, and footprint chips. Strict **non-duplication** of `AgentHeader` (it owns now+cost; Overview owns trend). Charts: `StackedBarChart.vue` (CSS/flexbox — not uPlot bars) + `TrendLineChart.vue` (uPlot, dark-mode-aware, custom cursor tooltip replacing the layout-shifting built-in legend); analogous-cool design-system palette. New agent-scoped analytics endpoint `GET /api/agents/{name}/analytics?window=` (`routers/analytics.py` → `db/schedules.py:get_agent_analytics`) generalises the #868 per-schedule query: `triggered_by` bucketing (+`Other` fallback), **full-set AVG vs sampled p95** correctness, terminal success rate, NULL-skipping context AVG, gap-filled UTC timeline. `executions.js` caches per `${name}:${window}`, never polled. `InfoPanel.vue` leads with About; `template.yaml` metadata behind a collapsible "Technical details". 12 unit tests. | [agent-overview-dashboard.md](feature-flows/agent-overview-dashboard.md), [agent-info-display.md](feature-flows/agent-info-display.md) |
 | 2026-06-06 | #1080 | feat: model-list refresh — **Claude Opus 4.8** (`claude-opus-4-8`) added as flagship across `ModelSelector.vue` PRESET + admin platform-default dropdown (`Settings.vue`); the two models retiring 2026-06-15 (`claude-opus-4-20250514`, `claude-sonnet-4-20250514`) removed from presets; current/legacy tiers relabeled. `TasksPanel.vue` localStorage fallback bumped legacy→`claude-sonnet-4-6`. MCP `model` param examples refreshed (`chat.ts`/`schedules.ts`/`loops.ts`). Backend/agent-server defaults already current (`claude-sonnet-4-6`, `claude-haiku-4-5-20251001`) — unchanged, no base-image rebuild. Graceful degradation: a removed preset stays valid free-text until Anthropic's retirement date, then fails with a clear execution error (not silently). Both model test files refreshed. | [model-selection.md](feature-flows/model-selection.md), [scheduling.md](feature-flows/scheduling.md) |
 | 2026-06-04 | VOIP-001 (#1056) | VoIP telephony Phase 1 (outbound) — an agent places a phone call over the **unmodified** Gemini Live bridge via Twilio Programmable Voice + Media Streams. New OSS three-layer module (`routers/voip.py` → `services/voip_service.py` → `db/voip.py`) + `voip_bindings`/`voip_call_logs` tables; a Media Streams WS adapter (`adapters/transports/twilio_media_stream.py` + pure `voip_audio.py`) with stateful `audioop` resampling (direct 8k↔16k/24k→8k), paced 160-byte μ-law sender, `clear`-on-barge-in; outbound trigger (REST + MCP `call_user`) with per-`(owner,dest)` rate limit + durable daily cap + Idempotency-Key; call-bound single-use WSS ticket (`ws_ticket_service` gains `ttl_seconds`); two-id namespace (`call_id` vs `vs_`); Redis staged-intent `GETDEL` consume-once; SETNX-guarded transcript save (`source="voice"`); **default-on post-call transcript processing** dispatched to the main agent (`execute_task(triggered_by="voip")`). Feature flag `voip_available` **OFF by default**. `voip_*` tables registered in the agent cascade registry. 27 unit tests; `audioop-lts` pinned for py≥3.13. | [voip-telephony.md](feature-flows/voip-telephony.md) |
